# swagger/swagger.yaml
swagger: '2.0'
info:
  description: Aplicación dirigida a compañías de Marketing para administrar todos los contactos de sus clientes para sus campañas.
  version: 1.0.0
  title: Data Warehouse
  contact:
    email: solmartinezpisani@gmail.com
...
schemes:
  - https
  - http
paths:
...
  /regiones:
    get: 
      tags:
        - regiones
      summary: Ver regiones
      operationId: getRegiones
      produces:
        - application/json
      parameters:
        - name: User token
          in: header
          required: true
          type: string
          description: User JWT token (Bearer)
      responses:
        '200':
          description: Successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/GetRegiones'
        '400':
          description: Unsuccessful
          schema:
            $ref: '#/definitions/Authenticate'
        '404':
          description: Not Found
          schema:
            type: object
            properties:
              status_code:
                type: integer
                format: int64
                example: 404
    post:
      tags:
        - regiones
      summary: Crear región
      operationId: createRegion
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: Crear región
          required: true
          schema:
            $ref: '#/definitions/CreateRegion'
      responses:
        '200':
          description: Successful Register
          schema:
            type: object
            properties:
              message:
                type: string
                example: Región agregada
        '400':
          description: Unsuccessful Register
          schema:
            type: object
            properties:
              status_code:
                type: integer
                format: int64
                example: 400
              message:
                type: string
                example: Región ya guardada
  /regiones/{id}:
    patch:
      tags:
        - regiones
      summary: Editar región
      operationId: updateRegion
      produces:
        - application/json
      parameters:
        - name: User token
          in: header
          required: true
          type: string
          description: User JWT token (Bearer)
        - in: path
          name: id
          type: integer
          required: true
          description: Numeric ID of the region to update.
        - in: body
          name: body
          description: >-
            Editar región
          required: true
          schema:
            $ref: '#/definitions/CreateRegion'
      responses:
        '200':
          description: Successful operation
          schema:
            properties:
              message:
                type: string
                example: Región actualizada
        '400':
          description: Unsuccessful
          schema:
            $ref: '#/definitions/Authenticate'
        '404':
          description: Not Found
          schema:
            type: object
            properties:
              status_code:
                type: integer
                format: int64
                example: 404
    delete: 
      tags:
        - regiones
      summary: Eliminar región
      operationId: deleteRegion
      produces:
        - application/json
      parameters:
        - name: User token
          in: header
          required: true
          type: string
          description: User JWT token (Bearer)
        - in: path
          name: id
          type: integer
          required: true
          description: Numeric ID of the region to delete.
      responses:
        '200':
          description: Successful operation
          schema:
            type: object
            properties:
              message:
                type: string
                example: Región eliminada
        '400':
          description: Unsuccessful
          schema:
            $ref: '#/definitions/Authenticate'
        '404':
          description: Not Found
          schema:
            type: object
            properties:
              status_code:
                type: integer
                format: int64
                example: 404     
  /paises: 
    post:
      tags:
        - paises
      summary: Crear país
      operationId: createPaís
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: Crear país
          required: true
          schema:
            $ref: '#/definitions/CreatePais'
      responses:
        '200':
          description: Successful Register
          schema:
            type: object
            properties:
              message:
                type: string
                example: País agregado
        '400':
          description: Unsuccessful Register
          schema:
            type: object
            properties:
              status_code:
                type: integer
                format: int64
                example: 400
              message:
                type: string
                example: País ya guardado
  /paises/{id}:
    patch:
      tags:
        - paises
      summary: Editar país
      operationId: updatePais
      produces:
        - application/json
      parameters:
        - name: User token
          in: header
          required: true
          type: string
          description: User JWT token (Bearer)
        - in: path
          name: id
          type: integer
          required: true
          description: Numeric ID of the country to update.
        - in: body
          name: body
          description: >-
            Editar país
          required: true
          schema:
            $ref: '#/definitions/CreatePais'
      responses:
        '200':
          description: Successful operation
          schema:
            properties:
              message:
                type: string
                example: País actualizado
        '400':
          description: Unsuccessful
          schema:
            $ref: '#/definitions/Authenticate'
        '404':
          description: Not Found
          schema:
            type: object
            properties:
              status_code:
                type: integer
                format: int64
                example: 404
    delete: 
      tags:
        - paises
      summary: Eliminar país
      operationId: deletePais
      produces:
        - application/json
      parameters:
        - name: User token
          in: header
          required: true
          type: string
          description: User JWT token (Bearer)
        - in: path
          name: id
          type: integer
          required: true
          description: Numeric ID of the country to delete.
      responses:
        '200':
          description: Successful operation
          schema:
            type: object
            properties:
              message:
                type: string
                example: País eliminado
        '400':
          description: Unsuccessful
          schema:
            $ref: '#/definitions/Authenticate'
        '404':
          description: Not Found
          schema:
            type: object
            properties:
              status_code:
                type: integer
                format: int64
                example: 404   
  /paises/regiones/{id}:
    get:
      tags:
        - paises
      summary: Ver un país de una región
      operationId: getPaises
      produces:
        - application/json
      parameters:
        - name: User token
          in: header
          required: true
          type: string
          description: User JWT token (Bearer)
        - in: path
          name: id
          type: integer
          required: true
          description: Numeric ID of the region that country belongs to.
      responses:
        '200':
          description: Successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/GetCountry'
        '400':
          description: Unsuccessful
          schema:
            $ref: '#/definitions/Authenticate'
        '404':
          description: Not Found
          schema:
            type: object
            properties:
              status_code:
                type: integer
                format: int64
                example: 404
  /ciudades:
    post:
      tags:
        - ciudades
      summary: Crear ciudad
      operationId: createCiudad
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: Crear ciudad
          required: true
          schema:
            $ref: '#/definitions/CreateCiudad'
      responses:
        '200':
          description: Successful Register
          schema:
            type: object
            properties:
              message:
                type: string
                example: Ciudad agregada
        '400':
          description: Unsuccessful Register
          schema:
            type: object
            properties:
              status_code:
                type: integer
                format: int64
                example: 400
              message:
                type: string
                example: Ciudad ya guardada
  /ciudades/{id}:
    patch:
      tags:
        - ciudades
      summary: Editar ciudad
      operationId: updateCiudad
      produces:
        - application/json
      parameters:
        - name: User token
          in: header
          required: true
          type: string
          description: User JWT token (Bearer)
        - in: path
          name: id
          type: integer
          required: true
          description: Numeric ID of the city to update.
        - in: body
          name: body
          description: >-
            Editar país
          required: true
          schema:
            $ref: '#/definitions/CreateCiudad'
      responses:
        '200':
          description: Successful operation
          schema:
            properties:
              message:
                type: string
                example: Ciudad actualizada
        '400':
          description: Unsuccessful
          schema:
            $ref: '#/definitions/Authenticate'
        '404':
          description: Not Found
          schema:
            type: object
            properties:
              status_code:
                type: integer
                format: int64
                example: 404
    delete: 
      tags:
        - ciudades
      summary: Eliminar ciudad
      operationId: deleteCiudad
      produces:
        - application/json
      parameters:
        - name: User token
          in: header
          required: true
          type: string
          description: User JWT token (Bearer)
        - in: path
          name: id
          type: integer
          required: true
          description: Numeric ID of the city to delete.
      responses:
        '200':
          description: Successful operation
          schema:
            type: object
            properties:
              message:
                type: string
                example: Ciudad eliminada
        '400':
          description: Unsuccessful
          schema:
            $ref: '#/definitions/Authenticate'
        '404':
          description: Not Found
          schema:
            type: object
            properties:
              status_code:
                type: integer
                format: int64
                example: 404   
  /ciudades/paises/{id}:
    get:
      tags:
        - ciudades
      summary: Ver una ciudad de un país
      operationId: getCiudades
      produces:
        - application/json
      parameters:
        - name: User token
          in: header
          required: true
          type: string
          description: User JWT token (Bearer)
        - in: path
          name: id
          type: integer
          required: true
          description: Numeric ID of the country that city belongs to.
      responses:
        '200':
          description: Successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/GetCity'
        '400':
          description: Unsuccessful
          schema:
            $ref: '#/definitions/Authenticate'
        '404':
          description: Not Found
          schema:
            type: object
            properties:
              status_code:
                type: integer
                format: int64
                example: 404
...
definitions:
  Authenticate:
    type: object
    properties:
      message:
        type: string
        example: Header must contain JWT to verify identity.
...
  GetRegiones:
    type: object
    properties:
      id:
        type: integer
        format: int64
        example: 1
      nombre:
        type: string
        example: Sudamérica
  GetCountry:
    type: object
    properties:
      id:
        type: integer
        format: int64
        example: 1
      region_id:
        type: integer
        format: int64
        example: 10
      nombre:
        type: string
        example: Argentina
  GetCity:
    type: object
    properties:
      id:
        type: integer
        format: int64
        example: 1
      pais_id:
        type: integer
        format: int64
        example: 10
      nombre:
        type: string
        example: Buenos Aires
...
  CreateRegion:
    type: object
    properties:
      nombre:
        type: string
        example: Sudamérica
  CreatePais:
    type: object
    properties:
      region_id:
        type: integer
        format: int64
        example: 10
      nombre:
        type: string
        example: Argentina
  CreateCiudad:
    type: object
    properties:
        pais_id:
          type: integer
          format: int64
          example: 3
        nombre:
          type: string
          example: Buenos Aires
...